% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{archer_fitN_odeint}
\alias{archer_fitN_odeint}
\title{Main optimization function for model with 5-8 parameters.}
\usage{
archer_fitN_odeint(
  parms,
  data,
  geno,
  I0 = NA_real_,
  pfCycleLength = NA_real_,
  inflec = NA_real_,
  ring_duration = NA_real_,
  circ_return = FALSE,
  seq_return = FALSE,
  ring_prop_return = FALSE,
  output_full_return = FALSE
)
}
\arguments{
\item{parms}{A numeric vector of parameter values.}

\item{data}{A dataframe containing necessary data (columns
\code{"Circ"} and \code{"ring_prop"}).}

\item{geno}{Decides which dataset is being optimized (if set as false,
then the function is fitting the data from O'Donnell et al., Parasite
Immunology, 2021; if set as true, then the function is fitting the data
from Prior et al., Scientifc Reports, 2019).}

\item{I0}{Single numeric indicating the initial population size.
Defaults to \code{NA}, which results in it being extracted from \code{parms}.}

\item{pfCycleLength}{Single numeric indicating the cycle length.
Defaults to \code{NA}, which results in it being extracted from \code{parms}.}

\item{inflec}{Single numeric indicating the inflection point.
Defaults to \code{NA}, which results in it being extracted from \code{parms}.}

\item{ring_duration}{Single numeric indicating the ring duration.
Defaults to \code{NA}, which results in it being extracted from \code{parms}.}

\item{circ_return}{Single logical indicating whether to output
circulating iRBCs.
Defaults to \code{FALSE}.}

\item{seq_return}{Single logical indicating whether to output
sequestered iRBCs.
Defaults to \code{FALSE}.}

\item{ring_prop_return}{Single logical indicating whether to output
ring proportions.
Defaults to \code{FALSE}.}

\item{output_full_return}{Single logical indicating whether to output
full ODE output.
Defaults to \code{FALSE}.}
}
\description{
Main optimization function for model with 5-8 parameters.
}
