% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_starts}
\alias{beta_starts}
\title{Produce a vector of starting abundances based on a symmetrical beta distribution}
\usage{
beta_starts(shape, offset, total0, compartments)
}
\arguments{
\item{shape}{Single numeric giving the Beta distribution shape parameter.}

\item{offset}{Single numeric giving the Beta distribution offset parameter.}

\item{total0}{Single numeric giving the total abundance across all stages.}

\item{compartments}{Integer giving the number of compartments for
stage structure.}
}
\description{
Produce a vector of starting abundances based on a symmetrical beta distribution
}
