% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constPMR_gammaN_ode}
\alias{constPMR_gammaN_ode}
\title{Run ODE for constant PMR and gamma N.}
\usage{
constPMR_gammaN_ode(x0, cycleLength, mu, museq, R, n, inflec, max_t, dt)
}
\arguments{
\item{x0}{Initial conditions for all stages.}

\item{Maximum}{time point to simulate to.}

\item{Time}{step to use for ODE.}
}
\value{
Numeric matrix where rows are time and columns are stages.
}
\description{
Run ODE for constant PMR and gamma N.
}
